unlzma: Fixes to get it to work in the init environment
API change: make the *posp argument return the decompressed data size
rather than the compressed size.

With the simple memory configuration present in the init environment,
the board hangs when trying to access values in the header struct when
it's packed (char + int + long). Work around this by reading the
individual bytes and bitwise-combining them together manually.

The firmware doesn't tell the stub the size of the kernel, but we don't
really need to know that. LONG_MAX (2^63 - 1) makes the decompressor
happy enough, it returns successfully when it hits the end of the
stream.